+++ kern/debug.c
@@ -72,15 +72,15 @@ debug_warn(const char *file, int line, const char *fmt,...)
 
 // Record the current call stack in eips[] by following the %ebp chain.
 void gcc_noinline
-debug_trace(uint64_t rbp, uint64_t eips[DEBUG_TRACEFRAMES])
+debug_trace(uint64_t rbp, uintptr_t eips[DEBUG_TRACEFRAMES])
 {
 #if SOL >= 1
-	const uint64_t *frame = (const uint64_t*)rbp;
+	const uintptr_t *frame = (const uintptr_t*)rbp;
 	int i;
 
 	for (i = 0; i < 10 && frame; i++) {
 		eips[i] = frame[1];		// saved %eip
-		frame = (uint64_t*)frame[0];	// saved rbp
+		frame = (uintptr_t*)frame[0];	// saved rbp
 	}
 	for (; i < 10; i++)	// zero out rest of eips
 		eips[i] = 0;
@@ -90,9 +90,9 @@ debug_trace(uint64_t rbp, uint64_t eips[DEBUG_TRACEFRAMES])
 }
 
 
-static void gcc_noinline f3(int r, uint64_t *e) { debug_trace(read_rbp(), e); }
-static void gcc_noinline f2(int r, uint64_t *e) { r & 2 ? f3(r,e) : f3(r,e); }
-static void gcc_noinline f1(int r, uint64_t *e) { r & 1 ? f2(r,e) : f2(r,e); }
+static void gcc_noinline f3(int r, uintptr_t *e) { debug_trace(read_rbp(), e); }
+static void gcc_noinline f2(int r, uintptr_t *e) { r & 2 ? f3(r,e) : f3(r,e); }
+static void gcc_noinline f1(int r, uintptr_t *e) { r & 1 ? f2(r,e) : f2(r,e); }
 
 // Test the backtrace implementation for correct operation
 void
